+++ src/lib.tsx
@@ -1,5 +1,7 @@
+import EasyCodeEditor from "./components/EasyCodeEditor";
+import DefaultLight from "./themes/DefaultLight";
+import DefaultDark from "./themes/DefaultDark";
 import "./index.css";
 
-export { default as EasyCodeEditor } from "./components/EasyCodeEditor";
-export { default as DefaultLight } from "./themes/DefaultLight";
-export { default as DefaultDark } from "./themes/DefaultDark";
+export { EasyCodeEditor, DefaultLight, DefaultDark };
+export default EasyCodeEditor;