+++ CHANGELOG.md
@@ -10,10 +10,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Simplified dev page
+- Default library export
 
 ### Changed
 
 - Optimised dynamic highlight rendering, performance is now tied to the supplied highlight function and the components textarea element
+- Simplified types
 
 ## [1.0.0] - 2024-07-20